requirements.txt
Flask==2.1.2
Keras==2.4.3
numpy==1.18.5
Pillow==9.1.1
